requirements.txt
asgiref==3.7.2
bokeh==3.3.1
celery==5.3.6
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
contourpy==1.2.0
cryptography==41.0.5
cycler==0.12.1
defusedxml==0.7.1
Django==4.2.6
django-allauth==0.58.1
django-appconf==1.0.6
django-compressor==4.4
django-filter==23.5
djangorestframework==3.14.0
fonttools==4.44.0
idna==3.4
Jinja2==3.1.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.1
numpy==1.26.2
oauthlib==3.2.2
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-decouple==3.8
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
rcssmin==1.1.1
redis==5.0.3
requests==2.31.0
requests-oauthlib==1.3.1
rjsmin==1.2.1
seaborn==0.13.0
six==1.16.0
sqlparse==0.4.4
tornado==6.3.3
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
xyzservices==2023.10.1
